kafka-assignment/requirements.txt
kafka-python
faker
python-dotenv
numpy
confluent-kafka
avro-python3
protobuf
grpcio-tools
requests